fix say, remove all stopwords from lunr
the search indexer lunr had a list of
stopwords that it ignored,
which is pretty useless for our usecase,
because we're not searching over
sentences, but over keywords,
which can include "say", "can", "what",
etc. - so in the best of cases this is useless and a waste of cpu cycles.