compound copy button highlights
clicking on the compound copy
to clipboard button now always
highlights the button.

this can be relevant when you search
for a compound, then deactivate
compound mode,
copy a single kanji,
re-enter compound mode
(at this point compound is not
highlighted) and
want to copy the compound again.
Before, it wouldn't be highlighted
again.