sync elements, search.js from gh-pages (#10)
sync with gh-pages 1.2.0.4

this is a big boy merge,
adding the elements from ~800
extra kanji annotated by elements
in the gh-pages branch.

this json was taken from visiting the webpage, finding the search.js in the sources, and applying a lot of regexes
to make the formatting the same as the old search.js, and the diff clean.