search() now returns results, has options, refactor
search returns a results object with
a length property and a property for
each kanji found,
and a list with ordered search results.

search function is now decoupled
from html elements, so you can
use the search function without a browser.

also moved the version information
in the offline version
from app.js to the index.html